requirements.txt
plotly
network
numpy
pandas
gunicorn

